`Benchmark` -> `Benchmark/HTTP`.
