+++ examples/Benchmark/HTTP/benchmark.cc
@@ -22,7 +22,7 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 *******************************************************************************/
 
-#include "../../current.h"
+#include "../../../current.h"
 
 #include "server.h"
